Work around parse errors of a bogus git.git commit
The C Git source code repository itself contains a commit with bogus
`gpgsig` lines: 5f549aa2f783 (pretty: %G[?GS] placeholders, 2011-10-21).
Basically, this commit's headers contain a GPG signature that _should_
have been a multi-line header, but instead every line of that signature
was added individually as a `gpgsig` header. Crucially, this includes a
header line that starts with `gpgsig` followed by a space and that's
already the end of the line.

Gitoxide's commit header parser was not prepared for that. Even worse:
the error message in that instance was simply empty.

Let's make the parser more robust by accepting such commit headers, and
add a test to verify that the commit in question _can_ be parsed by
Gitoxide.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>