fix(gix): extend lifetime of iterators
Previously the iterators would return references with lifetimes that
were shorter than the actual lifetimes of the `gix::Reference`
themselves. This was dues to a footgun with `'_` (eliding the lifetime).

When a function returns an elided lifetime, this lifetime usually is the
lifetime of the `&self` parameter, but sometimes it is the lifetime of
the type itself (e.g. `Iter<'_>`).

I made the lifetimes explicit to ensure we were using the correct ones.

Closes #2103