Update alternative locations integration tests for ARM64
This updates the integration-style tests for Windows-specific `git`
executable alternative locations to include `clangarm64/bin`
directories when looking under the 64-bit program files directory.
(These tests are "integration-style" in the sense that they check
the actual value of `ALTERNATIVE_LOCATIONS` against expectations
computed from details of the build target and runtime environment.)

The tests will now fail until `ALTERNATIVE_LOCATIONS` is fixed for
ARM64 Windows. (Because the currently intended approach is to check
for both ARM64 and x86_64 installations of Git for Windows on all
64-bit systems where the 64-bit "Program Files" directory location
is available -- regardless of whether the system is an x86_64 or
ARM64 installation of Windows -- these integration-style tests
should currently fail on both x86_64 and ARM64 Windows systems.)