Test that relative `EXEPATH` doesn't trigger the optimization
This tests that a nonempty value of `EXEPATH` that is a realtive
path does not trigger the `EXEPATH` optimization of
`system_prefix()` on Windows.

This is similar to the previously introduced test for an empty
value, in that both are treated as relative paths. Like that test,
this one currently fails, but both will pass once the `EXEPATH`
optimization is refined to refuse to look under a non-absolute
path.