Fix user program files alternative locations expectations
Unlike the global program files directories, the user program files
directory location (the location where it either already exists or
would be created the first time it is needed, when installing
software) is not architecture-specific. So we cannot decide which
of `clangarm64`, `mingw64`, and `mingw32` to look for inside of it
based on the values of environment variables, even with our build's
`target_pointer_width` to help. This is to say that `pf_user` is
fundamentally different from `pf_current`: while `pf_current` is
architecture-specific but in a way that varies by executable build,
`pf_user` is not architecture-specific.

This modifies the integration-style test expectation to include all
three subdirectories of a `Git` subdirectory of the user program
files directory, in all platforms.