Move the vacuous case outside the "ordinary" group
The vacuous case is ordinary, both in the specific relevant sense
of ordinary that there are no strange paths, and in the broader
sense that it is expected for overly sanitized environments
sometimes not to have any environment variables indicating where
program files directories are.

However, this is the simplest case, and calling it "ordinary" is
confusing. This moves it to be first, renames it, and removes the
now-obsolete comment.