Make `git remote` failure clearer in `make_remote_config_repos`
Starting in Git 2.51.0, at least one of the `git remote` commands
in the `make_remote_config_repos.sh` fixutre script fails, which
causes a subsequently run command depending on it to fail. This is
the cause of the CI failures observed in #2130.

But when any command fails, the intention is that the script fails
fast. To help achieve this, `set -e` is used. But the rules for
`set -e` have a number of exceptions. At least in some versions of
`bash`, the way `&&` was being used kept it from failing fast.

This changes that to run the commands separately, surfacing the
specific `git remote` command that fails, rather than a subsequent
command that fails later as a result of the first failure.