Use `git config` to make overlapping remotes in a test repo
Instead of `git remote`.

Since Git 2.51.0, path collisions in remote names are automatically
detected and rejected when `git remote add` is used to attempt to
add them, since creating such collisions is usually unintentional:
https://github.com/git/git/commit/a5a727c448e56b212546e047031166ced82c7d1e

However, we deliberately do this in our test suite, in the
`make_remote_config_repos.sh` fixture script. When Git 2.51.0 came
to be used on CI, some new test failures arose as a result, as
observed in #2130.

This fixes the script by using two `git config` commands rather
than a `git remote` command to produce the overlap. This technique
is (intentionally) still allowed in Git 2.51.0.