fix `gix-blame` performance regresion
The issue was introduced in https://github.com/GitoxideLabs/gitoxide/pull/1963, and incorrectly assumed that the `max-performance-safe` feature flag was only used to enable other feature flags. But this feature flag is itself used in various places, so failing to enable it leads to worse codegen. This issue was discussed/found in https://github.com/GitoxideLabs/gitoxide/issues/2148